Adds documentation about switchTo after	non-looping animation.
In the animation update	function, when an animation which does not
loop reaches its last frame, it	is no longer considered	playing.
This became confusing when following one of the	tutorials and finding
that the code example there does not work under	the current build.

This manifested	as a looping animation being stuck (not playing) on its
first frame after switching to it from a single frame non-looping animation.
The tutorial code at http://www.kiwijs.org/documentation/tutorials/creating-a-basic-kiwi-js-game/
does not reflect this as being expected	behavior.