Fix Animation frameIndex problems
Re issue #184

The last fix was not so good. It fixed problems with `onLoop` and `onComplete`, but it did it by avoiding `onLoop` and `onComplete` entirely.

The new fix employs mathematically rigorous methods to ensure that the `_frameIndex` property is within bounds before any external code is called. This should ensure that everything is fully functional at last.

Also rebuilt some docs.