Trash corrupt Textfield filenames.
The name `Textfield` is deprecated, as we changed it to `TextField` some versions ago. However, Git doesn't recognise case changes in filenames, so when the API documentation is rebuilt, it keeps the non-camel casing. This breaks links to `TextField` in the API documentation, as they are expecting camel case filenames, but the system does not rebuild them.

I'm hoping that by deleting those files with this commit, and later rebuilding them, the system will start defaulting to the camel case.