Fix `TextField` not recognizing arrays as color parameters.
Re GitHub issue #249

I've rearranged the constructor. It now passes `color` param to `color` property, which is an `Object.defineProperty` setter that correctly parses arrays.

The `vertexItems` declaration in `renderGL` is now in the location that it is used. This isn't hugely functional, but the code makes more sense now.

I've also reformatted the documentation, fixing the following errors:
- `render` did not identify its parameter
- `renderGL` did not describe its parameters