Fix `MouseCursor` losing track of inputs.
Re GitHub issue #259

Issue: Pressing multiple mouse buttons at the same time would cause Kiwi to lose track of which button was which when they went back up.

Cause: `MouseCursor` only sets `button`, `altKey`, `ctrlKey` and `shiftKey` in `start()`, which runs when the mouse cursor goes down. If a subsequent `start()` call is received, the information is overwritten and lost.

Solution: Enhance `MouseCursor.stop()` to update `button`. Add a `MouseCursor.move()` function extending `Pointer.move()`, in which the keys are updated. As `stop()` already invokes `move()`, this will ensure that the key data is updated very frequently. Although this is technically adding a function, `MouseCursor.move()` already existed on the prototype derived from `Pointer`. This is simply altering its functionality to fix a bug.

Result: `MouseCursor` now correctly returns the button of interest.