Implement `_validateChild()` to fix potential infinite loops in Group.
Re Github issue #266

Severe problems could occur if a cyclic dependency were created. For example, `getAllChildren()` would become an infinite loop. We now use a private method, `_validateChild()`, to prevent cycles in all `addChild` type methods.

Documentation was added to `Group.parent` to make it clear that this should not be set.