Add optional buffered substitute for tilemaps.
This system takes a series of screenshots of the tile layer, hides the original, and displays them as a series of stage-sized `StaticImage` objects. A component using a simple AABB keyhole limits visibility of these objects to keep rendering as fast as possible.

Analysis indicates that this can massively reduce the render time in a tile-heavy scene, but the initial screenshot process does take a human-significant amount of time (thanks to unpremulting), and low-end devices appear to struggle with the extra data; it's generally break-even, so I'm leaving it as an option.

More significant gains could be seen if we could take larger POT screenshots. I wanted to use a `GroupBuffer` camera aperture of 2048*2048, but this seemed to permanently screw up the state camera. We could be wasting a lot of VRAM with non-POT textures.