Accelerate rendering tiles in WebGL.
We don't have to do a matrix transform on every corner of every tile on the screen. This is easily 200 tiles, with 4 corners, for 800 matrix calculations (each 8 multiplications and 8 additions, plus function calls).

Instead, we can interpolate from one corner of the projected tile subgrid to the other, and use simpler computations. We replace 4 matrix operations (32*, 32+) with 12*, 12+ (or thereabouts) per cell. Also we're assigning lots of reused objects to local variables, reducing object hash lookups on `this.blah`.